Changes:
Modified Scene to include TextDisplay object, TextDisplayBehavior, and TextMesh child

Modified folder structure to allow for loading of Text asset

Imported a free monospace font

TextDisplayBehavior takes in a string (via SetText), which it divides into the correct number / size of lines to display, then displays it on the child TextMesh. Right now it's also loading in sample text from MobyDick; in the future, we can change this to use text from our book loading API. In the future, it would also be good to improve the way it divides text into lines, and add a feature to scroll text if it goes off the page. (it does not yet handle page overflow gracefully)