Client connects to server successfully.
Still need to do some work to make the client capable of listening for messages without freezing the GUI thread. Unfortunately, Unity doesn't support StreamReader.AsyncReadLine. It only supports ReadLine. The only workaround to this is to create a worker thread for the stream reader, then save all lines into a buffer of some kind, and the buffer must be locked. Then, when our ClientBehavior script updates, it will attempt to get the lock on the buffer, and one by one consume the buffer's lines with a delegate callback function. (LineRead(String str...)) Not sure whether ClientBehavior should read strings and call delegate, or read strings, parse JSON, and call delegate. Probably would be better to have a new AsyncStreamReader class, and have it call a delegate, and then ClientBehavior can be the delegate and consume the strings from JSON into objects, and then call a delegate function with those objects.